// frontend/lib/generated.ts
export const addresses = {
  hook: "0x7dEC15A39D42c9B5d41E0c351c0C9aDcbC8AC044",
  poolManager: "0x00B036B58a818B1BC34d502D3fE730Db729e62AC",
  swapRouter: "0x586E54DBEBd2999ef49f73632630AA8A689C6A6C",
  modifyLiquidityRouter: "0xDC37130a3f2D07EAacf6d8d66352932c79474b8a",
  currency0: "0x6BdACF5421d27aE8E88405d135e8A291F7B34B31",
  currency1: "0x818EB378DcAE64a4012c5bF0fDCC24c96fcBE57a",
} as const;

export const poolKey = {
  currency0: "0x6BdACF5421d27aE8E88405d135e8A291F7B34B31",
  currency1: "0x818EB378DcAE64a4012c5bF0fDCC24c96fcBE57a",
  fee: 500,
  tickSpacing: 60,
  hooks: "0x7dEC15A39D42c9B5d41E0c351c0C9aDcbC8AC044",
} as const;
// ...
export const swapRouterAbi = [
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_manager",
        "type": "address",
        "internalType": "contract IPoolManager"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "manager",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IPoolManager"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "swap",
    "inputs": [
      {
        "name": "key",
        "type": "tuple",
        "internalType": "struct PoolKey",
        "components": [
          {
            "name": "currency0",
            "type": "address",
            "internalType": "Currency"
          },
          {
            "name": "currency1",
            "type": "address",
            "internalType": "Currency"
          },
          {
            "name": "fee",
            "type": "uint24",
            "internalType": "uint24"
          },
          {
            "name": "tickSpacing",
            "type": "int24",
            "internalType": "int24"
          },
          {
            "name": "hooks",
            "type": "address",
            "internalType": "contract IHooks"
          }
        ]
      },
      {
        "name": "params",
        "type": "tuple",
        "internalType": "struct SwapParams",
        "components": [
          {
            "name": "zeroForOne",
            "type": "bool",
            "internalType": "bool"
          },
          {
            "name": "amountSpecified",
            "type": "int256",
            "internalType": "int256"
          },
          {
            "name": "sqrtPriceLimitX96",
            "type": "uint160",
            "internalType": "uint160"
          }
        ]
      },
      {
        "name": "testSettings",
        "type": "tuple",
        "internalType": "struct PoolSwapTest.TestSettings",
        "components": [
          {
            "name": "takeClaims",
            "type": "bool",
            "internalType": "bool"
          },
          {
            "name": "settleUsingBurn",
            "type": "bool",
            "internalType": "bool"
          }
        ]
      },
      {
        "name": "hookData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [
      {
        "name": "delta",
        "type": "int256",
        "internalType": "BalanceDelta"
      }
    ],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "unlockCallback",
    "inputs": [
      {
        "name": "rawData",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "error",
    "name": "NoSwapOccurred",
    "inputs": []
  }
] as const;